requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.1.7
djangorestframework==3.10.1
idna==2.8
PyMySQL==0.9.3
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
